requirements.txt
streamlit
pypdf 
sentence_transformers
sklearn.metrics.pairwise
skills_db